Add log rotation handling logic for mariadb audit logs
The existing log rotation logic assumes hourly rotation, but the MariaDB audit logs rotate based on log size. As a result, we can hang indefinitely once we reach the end of a log file.

More context on the PR here: https://github.com/honeycombio/rdslogs/pull/20

It looks like it's in some kind of loop where it never progresses the marker or reverts back to a 0 marker. I threw in debug logging on all code paths that may lead to this to get a better understanding of what's going on. I also added some logic to break the loop.